+++ RELEASE_NOTES.md
@@ -1,7 +1,9 @@
 # Release Notes
 
 ## 0.45
-- Default workflow inputs
+- Default workflow inputs. The `--defaults` command line argument
+embeds key-value pairs as workflow defaults. They can be overridden
+at runtime if necessary.
 
 ## 0.44
 - Use hashes instead of files for non-native dx types